1-CNN-Dogs-and-Cats-Classification/requirements.txt
numpy==2.3.4
Pillow==12.0.0
scikit_learn==1.7.2
streamlit==1.50.0
tensorflow==2.20.0
